backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.5
redis==5.0.4
celery==5.4.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
openai==1.30.1
tiktoken==0.7.0
cryptography==42.0.7
sentence-transformers==3.0.0
numpy==1.26.4
tenacity==8.3.0
structlog==24.1.0
python-dotenv==1.0.1
pytest==8.2.0
pytest-asyncio==0.23.6
